Incorrectly change the main branch, now restoring it

+++ src/DataTypes.java
@@ -7,7 +7,7 @@ public class DataTypes {
     }
     public static long sum(List<Integer> numbers) {
 
-        long s = 0;
+        int s = 0;
         // below is a "foreach" loop which iterates through numbers
         for (int x : numbers) {
             s += x;
